* Added banner art. * Fixed a bug in the MoveTo function that was caused when moving a file to the exact same location and deleting the original file when the overwrite flag is set.

+++ System.Files/File.cs
@@ -84,7 +84,7 @@ namespace System.Files
 		{
 			File PathFile = new File(Path);
 
-			if(Overwrite && PathFile.Exists)
+			if(Overwrite && PathFile.Exists && PathFile.FullName!=theFile.FullName)
 			{
 				PathFile.Delete();
 			}
